Refactoring a single function makes HTTP request
Refactored the code for downloading 'blob' as well as 'manifest' to use
a single function that performs HTTP request. This is like a first cut
implementaion. We are not yet accepting the 'body' to be sent while
making a request. And this function cannot be recursively called (for
redirects say!).