RwLock DockerClient: `bearer_token`/`auth_required`
There is only one function where Bearer token was getting updated. But
since that function was getting called from multiple places, we always
required `&mut` and this created problem. So now we protect the
`BearerToken` and `auth_required` inside `DockerClient`. So the callers
don't have to be mutable. This also makes calling `tags_list` with
checking bearer token, which was not possible before.