Added a method 'resolved_manifest' to Image trait
Actually, `Image` and `ImageSource` provide a lot of overlapping
functionality and we have copied these from the original Go
implementation. (Well not 'copied' but Used.) The current distinction is
as follows - `ImageSource` - Refers to the 'reference' (often user
provided) and `Image` - Refers to a 'resolved' one for the current OS
and Architecture. To get the manifest for the image we have defined a
new method `resolved_manifest` that corresponds to current OS/Arch.

This cleans up confusion.