Refactoring top level 'oci' module
Refactored Top Level 'oci' module to be part of the 'image' module. Most
of the functionality that is there in the 'oci' module right now is
related to image. (Cannot think of a use-case for Digest outside the
context of an image). This is perhaps not ideal, at some point we may
want to make 'oci types' only available as a part of 'crate' feature.
We'll see it then. Let's not worry so much about future requirements.