Using 'Stream' interface for getting Blobs
Initially, we were simply getting `Vec<u8>` for the blobs, this was okay
for small 'blobs' like Config blobs, but when downloading image layer
blobs, this interface is not appropriate as the blob will be huge
(running into MBs), so instead we are now using a 'Stream' interface
where we get `next` chunk in a loop, till we consume the requires
stream. Basic working implementation.