Added 'digest' output for 'inspect command'
We are now calculating the 'digest' output during the inspect command,
which was initially set to empty string. To be able to do this added a
`Digest::from_bytes` function to public API of `Digest`. This always
uses the default `sha256` algorithm.

Fixed a bunch of clippy warnings (`unwrap_or_else`)

In OCI Config PortString uses a `Hashmap<String, Empty>` now.