Getting started with 'mock' based tests for Image
Using 'wiremock' to 'mock' HTTP responses to add test cases for image
handling logic. We'll add expected responses (and/or errors) to test
different paths of image handling logic and Image APIs for the 'docker'
provider. Added a couple of simple tests to verify the reference, image
and the initial manifest.