Getting started with OCI Image Layout implementation
OCI Image Layout will be used serving two purposes - One for downloading
the images from a docker registry and storing them locally (something
like a `docker pull` functionality). Also, it should be possible to
create a local directory structure that implements this layout to be
used as OCI Image repository. The Image Layout will take care of
handling both the functionalities above, in addition, will be used as a
'source' for creating containers' `rootfs`.