Supporting image tags in `OCIImageLayout` and more
We are now accepting, optional image tag while creating a `new`
`OCIImageLayout` structure. The resultant layout will be namespaced
under the tag. While implementing this, cleaned up the layout directory
creation by making it a separate function which needs to be called
explicitly, thus `new` is now side-effect free (which is good!). Also
added a corresponding `delete_fs_path` function. Added a flag to test
whether the layout exists on the FS.