Fixed an issue with parsing references
If a reference name was given as `foo/bar:tag`, we were interpreting
this as domain('foo') and path('bar:tag'). However, a special handling
is required. First path is a domain only when it is a 'localhost' or a
proper domain in `a.b.c.d` format anything else should be treated as a
path component with domain set to default `docker.io`. This was observed
when trying to inspect the image 'docker://rustvmm/dev:v9' which was
failing with us but skopeo was 'inspect'ing it just fine. Also added
test cases for such references.