Using `AsyncRead` in place of `Stream` for blobs
For all the blobs while verifying digests etc, we were using `Stream`
interface. Moved that interface to `AsyncRead`, which is a lot more
intuitive, plus most of the times when we'll be dealing with 'blobs'
they will be on a local file system (while verifying, gunziping etc.)
rather from a network stream, so `AsyncRead` is a better interface to
work with. Also, for using `GzipDecoder` it is not possible to use
`Stream` interface yet (with current async-compression version 0.3.7),
due to some version incompatibility. Thus `AsyncRead` is better here as
well.