Fixed hard-coded Layer Digest in overlay test
While testing `apply_layer` for the overlay storage, we were using
hard-coded digest. However this digest would change, even for the same
tag and that would cause the test for `apply_layer` to fail. Since we
are anyways 'pull'ing the image, we parse the image manifest and get the
actual layer digest and use it now. This means some boiler-plate code in
the test, but for now okay.